+++ changelog.md
@@ -1,5 +1,6 @@
 # Version 20.10
 
++ Add "VPoint.is_slider" method.
 + Improve performance of meta-heuristic algorithms.
     + Release almost GIL, only left objective function.